Add filter terms search in search suggestions and remove input value for using category filter
